Improvements to rpc-tests.py (add -force-enable / -f , better options handling)
This is a cross-port of improvements to the test runner from MVF-BU.

It adds a -f option to force running of specified disabled tests.
It also passes down test-specific option correctly.

The previous command line option handling was severely broken - it still is not great, but this improves it to the point where it is possible to run e.g.

$ qa/pull-tester/rpc-tests.py excessive --extensive

Because the 'excessive.py' test used the '--extended' option which is now classed as a 'private' option of the top-level runner script (and therefore does not get passed on to the called script), that option had to be renamed to '--extensive'.

Any tests that take their own options should have these added in the `test_script_opts` list in rpc.tests.py .

I've refrained from reformatting rpc-tests.py using PEP8 as I'll do a general overhaul of the Python scripts in a later PR.