Count bytes as we build a thinblock
And clear thinblock data and disconnect a node if the maximum bytes
permitted for concurrent thinblock re-construction has been exceeded.

If only one thinblock is currently being processed then it will
be the one selected for termination.

If more than one thinblock is in flight then the largest thinblock
will be terminated and the node disconnected.

Disconnecting the largest thinblock will automatically clear the
the thinblock data.