Factor out information unit formatting
There's a couple places where a floating point amount of bytes is
formatted  with units attached (KB, MB, ...).

This factors all of them out and puts them into a single function
string formatInfoUnit(double).

It also fixes a couple potential array-out-of-bounds-access errors,
as none of the upper bounds had been checked before, and adds docs
and a unit test.