Partial clean up of copy & paste code for xthin stats
There's a lot of repeated code to deal with expiry of old
statistics values after a day.

This commit replaces the repeated sections with a call to a
template function to update the statistics and expire old
values.

It also replaces repeated average calculation code with a
method average().

Includes (partial) test cases for the CThinBlockData structure.
(Missing those parts that depend on chain state)