Use a bool return value for RequestBlock
This fixes a bug in the Request Manager when the first block source
can not be downloaded from because it is not a XTHIN peer.  Previously
we would have to wait for the download timeout to be exceeded before
trying again which could be as much as 30 seconds if rate limiting
was turned on.  Now as soon as another block source is available it
will be checked until we make our first request.