stat.h: Fix missing inits and wrong type init
This fixes:

  - value being uninitialized when using the default constructor
    and fundamental data type as RecordType in CStat.
  - op and sampleCount being uninitialized in CStatHistory
  - total in CStatHistory was initialized with DataType() constructor,
     but actually is a RecordType.

It also initializes op(eration) in CStatHistory with
STAT_OP_SUM and STAT_KEEP_COUNT now.

STAT_OP_SUM was chosen as this is the default of the other constructors.

Addition of the STAT_KEEP_COUNT flag (and only this part) fixes issue #614
for me.

It appears as if timeout(..) reset that flag before and memory was
by chance always (un)initialized to have that flag set in op.