Move bitnodes seeding to be after DNS seeding.
By removing the bitnodes seeding thread and moving the bitnodes
seeding to run after DNS seeding we prevent the problem on first
startup where binodes seeding finishes first and we end up with
100 entries from the leaderboard at the beginning of our address
list.  Whereas what we wanted is for the first entries to be coming
from our on BU seed list. And if no seeds could be found then and
only then get them from the bitnodes leader board.