Always add every block source during IBD
In the past we were only adding a block source if there was no
block currently in flight.  We want to always add more block
sources in the event that a block in flight is delayed. This way
we give the request manager the ability to re-request the block
from another source.

This is a possible bug fix for issue #575