Do not assign Misbehavior if unconnecting headers are received
This is typcially not a misbehavior and can and does happen on
mainnet when two blocks are mined by different miners but very
close in time to one another.  Some nodes will receive these
headers out of order and while on mainnet giving a small misbehavior
is not a big issue for this rare event it has caused some problems
on the nolnet test environment.

The next step will be to add the ability to cache these un-connecting
headers until such time as the needed header arrives and both can
then be properly connected.