Handle headers that won't connect right away.
At times we can receive out of order headers.  If we do we store
the unconnecting header in a cache and check if it will connect
each time a new header arrives.

Out of order headers should typically
exist for only a few seconds at most, therefore if the unconnecting
header is in cache for more than 2 minutes we will delete it.

There is also the possiblity of an attack to prevent node sync from
happening during the intial download of headers.  If we get out of
order headers and we are far from today then we disconnect the peer
so that we can start downloading initial headers from another peer.