Performance improvement for in memory dbcache
When the cache is periodically flushed we don't want to
just clear the entire cache because there are still many
good entries there which then have to be pulled from disk
again if needed.  Rather, we wan't to just prune the entries
that are DIRTY which need to be flushed to disk and keep any
unmodified coins which can be used later during block
validation.

Add logprint for number of Trimmed entries