Continuously trim the dbcache while preventing continuous flushing
We want to make sure the cache is trimmed to it's max configured size
however we only want to do a full flush when needed which is usually
after a block is processed.  After a block is processed we check to
see if we are close to the max configured cache size and then proceed
to flush only when we're near.