Also check pindexBestHeader when determining fScriptChecks
Since we can be reasonable sure we have all the headers when we
do IBD or receive new blocks then we can also check if we are
within approx 30 days of the chain tip by block height.  We still
keep the check that timeBarrier is also within 30 days but by
having both checks it makes it more difficult to attack in some
way.