Move block relay to ActivateBestChainStep
In PV we can connect blocks beyone pindexMostWork if other
blocks arrive just after we start validation.  Therefore we
need to relay block inventory based on the new pindexMostWork
after all the blocks have been connected.  If we instead
return to ActivateBestChain we will relay inventory based on the
old pindexMostWork and may end up not relaying a block.