Check for fRequestShutdown in ConnectTip()
Move the check for a requested shutdown from ActivateBestChainStep()
into ConnectTip().  This makes sure that whenever and wherever
we try to connect another block we make the check. In PV sometimes
we will loop and try to connect many blocks at once time before
returning so this way we can be sure the check is always being made
for each block.