Use `os.path.join` rather than string concatenation
This would guarantee to generate files path that
are valid in every kind of OS you are going to run
the tests.

Unfortunately it seems there're a lot of other places
where string concatenation is used. I think we should
fix all of them if we want the tests suite to be portable.