Migrate QT settings for BUcash nodes
In the case where settings were just migrated as part of startup, if
this is a BUcash client, force a re-confirmation of the data directory,
which will be stored in the registry (Windows) or a settings file
(Linux/Mac).  This is because if the user wants to run dual nodes for
BTC and BCC chains on the same machine, it will require separate data
directories.  This only affects bitcoin-qt.