Update UndoReadFromDisk
This was missing after all the cherry-pick's for the database
upgrade. The code for Core is in Validation.cpp but for BU it
is in main.cpp.  During these kind of cherry-picks the code
has to be manually carried over however sometime with so many
changes one goes missing. In the future we should better align
the BU codebase with Core so that these type of cherry-pick
issues don't needlessly show up.