Fix spurious failures in parallel.py
The test was occasionally failing due to a timing issue with
the largest block of a two block race arriving too quickly and
being validated before the small block would show up.  This was
resulting in an assertion being thrown indicating that the small
block did not win the race, when in fact, a race hadn't even
occurred.  Here we set -pvtest=1 to make sure that the validation
of the inputs is done with 1 second thinktimes in order to slow the
validations down and make sure that both blocks arrive before
any block can finish it's own validation.

a chance to have a parallel validation race.