Fix spurious failures in coins_tests.cpp (#865)
* Fix spurious failures in coins_tests.cpp

Recently we reduced the number of iterations this test runs
from 40000 to 10000. While this greatly improved the time it
takes for tests to finish on Travis it introduced a random
failure in one of the tests because it was not getting executed
often enough to hit one of the tests, which is exectued also
by random chance.  By running that test more frequently the problem
should be solved.

* Tidy up a few comments in coins_tests.cpp