Trim coin cache by coin height (#812)
* Introduce nBestCoinHeight and track it's value in each coin cache

* Have SaltedOutPointHasher return a uint64_t

We no longer use a boost map therefore we can return a uin64_t
on all systems rather than size_t.  This was previously an issue
on a narrow range of boost versions which resulted in failures when syncing
the chain (Core issue #4634).

* Trim coins cache by coin height

The cache is trimmed by the coin height.  If the height is not
high enough to trim any values then we re-adjust the trim height
up towards the current chain tip until we trim enough coins. If
there are still not enough coins to be trimmed then we just trim
any coins than are not dirty.

Futhermore, once the coin height has moved closer to the tip and
we end up not needing to trim any coins when trim is called, then
we gradually walk the trim height away from the tip so when trim
is once again succesful we will potentially maintain more of the
recent in cache.

This process of trimming by height only comes into effect when
the cache limit is exceeded and is therefore only used by more
marginal nodes that do not have a large dbcache.

* formatting

* Remove unnecessary iterator definition

* Simplify the calculation of nTrimHeightDelta

* Make nSmallestDelta a uint64_t to fix compiler warning

* Use saturating arithmetic for nTrimHeightDelta and nTrimHeight

* Add Trim test to coin_tests.cpp

Every once in a while trim a random cache to zero. This tests that
we never get into any kind of a spin loop.