Bug Fix for Coin Freeze when making payment request (#875)
Previously when making a new payment request, that would
generate a new P2SH freeze address, there would be another
P2KH address added to the address book which would also
have the same label applied.  This second address is not
needed and is confusing to the user who was only trying
to create only one freeze address.  The fix here is to only
create the P2SH freeze address as was intended.