Add init message for Reaccepting Wallet Transactions (#877)
During startup the very last step is to reaccept wallet transactions.
Previously the client was indicating that loading was complete before
this last step. Here we send a ui message and also print to the logfile
to indicate that wallet transactions are still being reacceted and then
only when finally complete we indicate that loading is finished.