Add helper functions to implement time based fork
If 'x-1' is used to identify the first block for which GetMedianTimePast()
(GMTP) is equal or greater than fork time and 'x' to indicate the first
block mined after the time based trigger fired. The following helper will
check if a given block belongs to 4 different intervals, namely:

- [x,+inf)
- [x,x]
- [x-1,+inf)
- [x-1,x-1]

A fork is considered to be enabled at height x-1 and activated at height x.
We chose this naming scheme because usually block 'x' has to satisfy
additional conditions (e.g. block size in the UAHF case)