Temporarily disable bitnodes seeding on the BitcoinCash chain
There is no bitnodes service for BitcoinCash and therefore
when this section of code is run we end up pulling into  our
peers.dat a list peers that are valid for the legacy chain
and not the BitcoinCash chain. Once a simlilar site is
available we can easily re-enable this feature.