New RPC feature : rollbackchain (#931)
* Create rpc to roll back the blockchain to any height

This is a hidden feature which can be used during testing
or by users that are having difficulty syncing to the correct
chain and have tried reconsiderblock but have failed. By rolling
back beyond the fork point a user can then invalidate and reconsider
the correct chain.

* Fix bug in reconsiderblock

We need to make sure to update the ui after we reconsider. This
was not always happening if we rolled the chain back and then reconsidered
a block again.

* Remove unnecessary spaces

* Limit how far a rollbackchain can go

In case of operator error, limit the rollback to 100 blocks
at a time.  An operator may for instance want to roll back
10 blocks and then enter "rollbackchain 10" which would have
the effect of rolling the chain all the way to block height 10
rather than just 10 blocks.

* Do no resurrect transactions when doing rollbackchain

Resurrecting transactions is not really needed since the operator
may be following a different chaind with different txns and also
it slows down the process of rollback.