Continue to the next node after getting a socket error (#933)
* Continue to the next node after getting a socket error

Do not attempt to read any more bytes from a socket that had
and error and has been issued an fDisconnect.

* If we have been issued an fDisconnect then don't try to send any messages

It's possible that the socket associated with this node had an error
when bytes were being read in and was issued an fDisconect. If true
there is no need to try to send messages out on this connection.