Automatically determine a more optimal -dbcache setting if none provided (#948)
* Kill QT dbcache setting in Options

This setting is of little use since the operator has to restart
the application anyway to make the setting take effect.  By removing
this and only using -dbcache from either the command line or bitcoin.conf
it gives us the ability to automatically configure the dbcache and
also simplifies the configuration process.

* Add functions to determine the max system memory and use them

Use half the total system memory as the value for -dbcache
and use this value if the node operator has not already set
a nummber for -dbcache.  This way we can be pretty sure a node
will get a much better initial sync rather than in the past where
we would only rely on the default of 500MB for -dbcache.

Add getTotalSystemMemory() for UNIX and MACOS

Also return with a default value if physical memory size could
not be determined. We return with a default of 1GB which should
cover any raspberry pi configurations.

Make a more accurate calculation on Windows for memory available

On Windows systems we can know how much of the physical RAM is available
for use and so here we make a more fine grained calculation for the
amount of memory we'll use for the dbcache mechanism. For Windows systems
with large amounts of RAM available we'll use all of it up to the 16GB
maximum allowed for dbcaching, however, we'll always leave 1GB free
to the system for other uses.

* Use an int64_t instead of size_t for nCoinCacheUsage

Because nCoinCacheUsage is size_t, then on 32bit systems when
we assign it a value from 64bit int then we end up with an
incorrect value. This happens during init when the cache sizes
are being calculated.

* Move memory functions to memusage.h in preparation for dynamic dbcache sizing on Windows

* On Windows, dynamically reduce the size of the dbcache if/when needed.

If the available memory falls below 10% then
reduce the size of the dbcache's coin cache.

* Allow the retreival of the lowest default dbcache configuration

This is needed to prevent the coins cache from getting too low
during the process of dynamically changing the dbcache settings.

* On Windows, dynamically increase the size of the dbcache if possible.

* Cleanup FlushStateToDisk by moving the dynamic coin cache adjustment

Create AdjustCoinCacheSize() and use it in FlushStateToDisk and move
the underlying code to txdb.cpp.

* change from LOGA to LOG(COINDB, ...

* Update README.md with instruction on setting up for initial sync.

Describe the importance of the dbcache setting in the initial sync
process. Also touch upon the automatic dbcache configuration as well
as the possible need to set a higher maxoutconnections.

* Add comments for thew new functions in txdb.h

* Fix GetTotalSystemMemory for MacOS

a zero indicates a positive return value for sysctl rather
than a 1

* Use a better size estimate when determining when to batch write

Since the time of the datbase upgrade the batch size was not
being calculated properly and as a result we were not batching
our writes anymore. Here we get the appropriate batch size
estimate and write our batches accordingly.

* On Windows, ensure that there is enough free memory for the write caches

On very marginal nodes we have to at least leave enough memory for the
write caches and also be careful to trim the coins cache more
carefully.

* fix compiler warnings

* Once IBD is complete then trim the coins cache

If the user has not specifiied a dbcache setting then trim
the coins cache to the default size once IBD is completed.

* Flush state and trim coins cache before shutting down

At the beginning of the shutdown process, flush and clear the coins
cache to free up as much memory as possible. This helps to prevent
any runaway exceptions due to a bad alloc during the shutdown process.

* Keep 300MB available for marginal Windows systems