Remove unnecessary code for retrieving headers
In the past we had issues on regtest when retriving blocks based on
INV's. Blocks could return out of order which then led to them getting
rejected because the header chain was not up to date. To get around this,
on regtest, we would ask for the headers again if the chain was not
up to date, however, we no longer request blocks by INV. Instead if
we receive an INV for a block we automatially request the header first
and only on receiving and updating a valid header do we then ask for
the block. Thus, this regtest workaround is no longer needed.