Properly check whether a transaction was previously in a block or not (#1080)
* Track recent transactions in blocks and do not re-request them

*Thanks to Matt Corallo (Sorry, I wasn't able to do a proper cherry
pick here so couldn't give proper attibution.)

It is possible that transaction announcements arrive after a block
is processed and we do not want to re-request those txns. In the past
we used the existence of coins in the coins cache as evidence of these
transactions however in BU those coins will likely no longer be in cache
anyway once the block is handled, due to the way we trim the dbcache,
and furthermore we would have required a lock on the utxo to be
maintained (which we were in fact missing).  This approach relies
on a separate lock on cs_recentRejects which avoids any contention
and also ensures we don't ask for those txns again.

* Simplify the logic in AlreadyHave and make it more readable

* Add DbgAssert to recentRejects in AlreadyHave

* Split AlreadyHave() into AlreadyHaveTx() and AlreadyHaveBlock()

This is easier to maintain and prevents us from requiring cs_main
for AlreadyHaveTx().

* Reduce scope of cs_main when we receive INV messages

We only have to lock cs_main if we received block inventory, which
is in fact rare now.  Most INV's are for transactions and we can
save a great deal of cs_main locking here.

* Reduce scope of cs_recentRejects

Each lock in AlreadyHaveTx is now taken outside the scope of any of
the other locks so that we don't have to worry now or in the future
about locking orders.

* swap boost scoped_ptr for std::unique_ptr

* Use excessive block size to calculate the number of bloom filter elements

For transactions recently in a block we want to automatically update
the number of elements in the bloom filter to be dependant on the size
of EB. Although we could in theory receive a larger block this calculation
should be sufficent and it does not have to be 100% accurate. The worst
case being we end up asking for a few transations that we already have.
But in general the filtering done here should prevent all transactions
from being asked for and downloaded twice.